+++ data/data.txt
@@ -11,8 +11,7 @@ t'es aussi sécurisant qu'une segmentation fault
 t'es aussi chiant qu'un virus, sauf que lui il est codé pour ça
 t'as pas assez de neurones pour faire du multithread
 t'as redoublé ta crèche tellement t'es pas adapté à l'intelligence
-t'as été codé en hook
-t'es aussi clair qu'INTERCAL
+t'as été codé en ook
 t'as breaké une boucle alors que t'étais même pas dedans. C'est con, c'était celle qui relançait es instructions d'intelligence
 t'as arrêté ton développement à la prophase
 t'as raté le carrefour phylogénique qui sépare le singe de l'Homme
@@ -28,8 +27,17 @@ ton pc n'a jamais vu le terminal, ni même son émulateur
 ton pc tourne sous brainfuck OS
 ta méthode parler() n'a pas d'argument
 ta carte son ne sait faire que du bruit chiant
+je te chie dans la bouche
+je te vomit sur le visage
+je motoculte ta maman tout les matins
+je te pisse à la raie
 t'es qu'une espèce de moule asmathique
-toi t'as une gueule à tourner sous Windows,
+toi t'as une gueule à tourner sous Windows
+ton papa n'est pas gentil mon cher
+je déchire l'anus de ta soeur
+que je vais te ramoner à sec avec du gravier
 t'es un faquin
 le mot beauté est mieux adapté à Susan Boyle qu'à toi
-le train de mes insultes roules sur les rails de l'indifférence de
+j'espère que tu chopperas rapidement une blue waffle
+le train de mes insultes roules sur les rails de ton indifférence
+je t'aime plus que tout et que j'ai secrètement envie de te faire des enfants
