+++ data/whitelist.txt
@@ -1,0 +1,3 @@
+plopp
+aluriak
+neolem
